Use role=heading and aria-level inside disclosures, as a dumb workaround for some Chrome-related issues with the details element
Specifically: putting an h2 inside a summary seems to cause the decorative triangle icon to not be hidden from screen reader users.  The icon can be hidden with CSS, but this gets rid of it for all users and also means that the expanded state is never conveyed to screen readers.